split code into smaller files. converted project into a crate. implemented DiskEntry struct for better file creation
