+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod app;
+pub mod result;
+pub mod args;
+pub mod file_manager;
